requirements.txt
SQLAlchemy
python-dotenv
psycopg2
fastapi
numpy
schedule
pytest
almebic